Multiple changes
Added comments to parent class and subclasses for documentation. Changed health, damage, and directDamage attributes to int for consistency and to not need typecasting. Fixed logic and syntax in action() method in subclass Cherrybomb to access Zombie attributes through methods instead of directly. Fixed syntactical and logical errors. Scaled the attribute values of each plant to the attribute values of zombies